+++ Client/client/Entities/Player.cpp
@@ -1,7 +1,7 @@
 #include "Player.hpp"
 
 
-Player::Player(int32 p_ID, std::string p_Name, uint8 p_Level, uint8 p_Health, uint8 p_Mana, uint8 p_Alignment, uint8 p_SkinID, uint8 p_SizeX, uint8 p_SizeY, uint16 p_MapID, uint32 p_PosX, uint32 p_PosY, Orientation p_Orientation) :
+Player::Player(int32 p_ID, std::string p_Name, uint8 p_Level, uint8 p_Health, uint8 p_Mana, uint8 p_Alignment, int16 p_SkinID, uint8 p_SizeX, uint8 p_SizeY, uint16 p_MapID, uint32 p_PosX, uint32 p_PosY, Orientation p_Orientation) :
     Unit(p_ID, TypeUnit::PLAYER, p_SizeX, p_SizeY)
 {
     m_Resources[eResourceType::Mana] = new Resource(eResourceType::Mana);
