+++ Server/server/Entities/Unit.cpp
@@ -501,7 +501,7 @@ void Unit::AddResourceNb(eResourceType p_Resource, uint8 p_Nb)
     }
 }
 
-uint8 Unit::GetSkinID() const
+int16 Unit::GetSkinID() const
 {
     return m_SkinID;
 }
@@ -564,7 +564,7 @@ void Unit::SetLevel(const uint8 & p_Level)
     m_Level = p_Level;
 }
 
-void Unit::SetSkinID(const uint8 & p_SkinID)
+void Unit::SetSkinID(const int16 & p_SkinID)
 {
     m_SkinID = p_SkinID;
